+++ jvm/sargon-desktop-bins/.gitignore
@@ -1,0 +1,2 @@
+/build
+/src/main/resources/
